chore: code cleanup sweep
Dedupe resolveHostUnheadMajor by consuming nuxtseo-shared/kit's shipped
version instead of the local copy (TODO'd for this exact migration),
extract the duplicated content:file:afterParse script-building logic
(module.ts + nuxt-content-v2.ts) into a shared buildSchemaOrgContentScript
util, and collapse three identical useHead branches in useSchemaOrg into
one condition.
