+++ src/kit.ts
@@ -1,1 +1,2 @@
-export { extendTypes, resolveNuxtContentVersion } from 'nuxtseo-shared/kit'
+export { extendTypes, resolveHostUnheadMajor, resolveNuxtContentVersion } from 'nuxtseo-shared/kit'
+export type { UnheadMajor } from 'nuxtseo-shared/kit'
