+++ src/unhead-compat.ts
@@ -1,9 +1,7 @@
+import type { UnheadMajor } from './kit'
 import { existsSync } from 'node:fs'
-import { dirname } from 'node:path'
-import { pathToFileURL } from 'node:url'
-import { readPackageJSON, resolvePackageJSON } from 'pkg-types'
 
-export type UnheadMajor = 2 | 3
+export type { UnheadMajor } from './kit'
 
 function isPlainObject(value: unknown): value is Record<string, any> {
   return !!value && typeof value === 'object' && !Array.isArray(value)
@@ -49,65 +47,6 @@ export function resolveSerializableIdentityConfig<T>(identity: T): T {
     cloned.type = resolvedType
 
   return cloned as T
-}
-
-function normalizePackageUrl(rootDir: string): string {
-  const url = rootDir.startsWith('file:')
-    ? rootDir
-    : pathToFileURL(rootDir.endsWith('/') ? rootDir : `${rootDir}/`).href
-  return url.endsWith('/') ? url : `${url}/`
-}
-
-/**
- * `@unhead/schema-org` attaches an object `_resolver` to every graph node. Unhead
- * v3's `walkResolver` skips that key, but v2's walks into it and invokes the
- * resolver's `resolve()` method as an argument-less thunk, crashing with
- * "Cannot read properties of undefined (reading 'potentialAction')" (#114).
- *
- * Empirically schema-org v2 renders cleanly on unhead v2, and schema-org v3 on
- * unhead v3. So we ship both majors (`@unhead/schema-org` + the aliased
- * `@unhead/schema-org-v2`) and select the one matching the host's unhead.
- *
- * Returns the major of the unhead the host app renders with, falling back to the
- * module's primary major (3) when it can't be resolved.
- *
- * TODO: once nuxtseo-shared ships `resolveHostUnheadMajor` (nuxt-seo#561), import
- * it from `nuxtseo-shared/kit` and delete this local copy.
- */
-export async function resolveHostUnheadMajor(rootDir: string): Promise<UnheadMajor> {
-  const rootUrl = normalizePackageUrl(rootDir)
-  // Search from the packages that own SSR before the app root. A host can have
-  // @unhead/vue v3 installed at the project root while Nuxt/Nitro renders with
-  // nested @unhead/vue v2; matching the root copy would still crash SSR (#114).
-  const searchUrls = []
-  for (const id of ['@nuxt/nitro-server', 'nuxt']) {
-    const pkgJson = await resolvePackageJSON(id, { url: rootUrl }).catch(() => {
-      // A missing package is an expected miss in non-standard layouts; keep
-      // searching from the remaining candidates.
-      return undefined
-    })
-    if (pkgJson)
-      searchUrls.push(`${dirname(pkgJson)}/`)
-  }
-  searchUrls.push(rootUrl)
-  // `@unhead/vue` is what schema-org actually peer-depends on; fall back to the
-  // core `unhead` package when it isn't directly resolvable.
-  for (const id of ['@unhead/vue', 'unhead']) {
-    for (const url of searchUrls) {
-      const version = await readPackageJSON(id, { url }).then(pkg => pkg.version).catch(() => {
-        // an unresolvable package is an expected miss (not installed under this id
-        // at this search root); fall through to the next candidate/root, then the
-        // default major.
-        return undefined
-      })
-      const major = version ? Number.parseInt(version, 10) : Number.NaN
-      if (major === 2)
-        return 2
-      if (Number.isFinite(major) && major >= 3)
-        return 3
-    }
-  }
-  return 3
 }
 
 export type SchemaOrgVendor
